# openapi.json
{"openapi":"3.0.2","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/wordseg":{"post":{"summary":"Wordseg","operationId":"wordseg_wordseg_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_wordseg_wordseg_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/daily":{"get":{"summary":"Guess Daily","description":"Guess against the daily puzzle","operationId":"guess_daily_daily_get","parameters":[{"required":true,"schema":{"title":"Guess","type":"string"},"name":"guess","in":"query"},{"required":false,"schema":{"title":"Size","type":"integer","default":5},"name":"size","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Guess Daily Daily Get","type":"array","items":{"$ref":"#/components/schemas/GuessResult"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/random":{"get":{"summary":"Guess Random","description":"Guess against a random word","operationId":"guess_random_random_get","parameters":[{"required":true,"schema":{"title":"Guess","type":"string"},"name":"guess","in":"query"},{"required":false,"schema":{"title":"Size","type":"integer","default":5},"name":"size","in":"query"},{"required":false,"schema":{"title":"Seed","type":"integer"},"name":"seed","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Guess Random Random Get","type":"array","items":{"$ref":"#/components/schemas/GuessResult"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/word/{word}":{"get":{"summary":"Guess Word","description":"Guess against a selected word","operationId":"guess_word_word__word__get","parameters":[{"required":true,"schema":{"title":"Word","type":"string"},"name":"word","in":"path"},{"required":true,"schema":{"title":"Guess","type":"string"},"name":"guess","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Guess Word Word  Word  Get","type":"array","items":{"$ref":"#/components/schemas/GuessResult"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_wordseg_wordseg_post":{"title":"Body_wordseg_wordseg_post","required":["text"],"type":"object","properties":{"text":{"title":"Text","type":"string"}}},"GuessResult":{"title":"GuessResult","required":["slot","guess","result"],"type":"object","properties":{"slot":{"title":"Slot","type":"integer"},"guess":{"title":"Guess","type":"string"},"result":{"$ref":"#/components/schemas/ResultKind"}}},"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"ResultKind":{"title":"ResultKind","enum":["absent","present","correct"],"type":"string","description":"An enumeration."},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"type":"string"}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}}}}